# artifacts/api-server/src/routes/me/keep.ts
import { Router, type IRouter } from "express";
import {
  db,
  serviceConnectionsTable,
  libraryItemsTable,
  keepTargetsTable,
  pendingKeepsTable,
  recordingsTable,
  spinsTable,
  stationsTable,
  showsTable,
  pickersTable,
  supportHoldsTable,
  type LibraryItemProvenance,
} from "@workspace/db";
import {
  HoldRecordingSupportParams,
  HoldRecordingSupportResponse,
  UnholdRecordingSupportParams,
  UnholdRecordingSupportResponse,
} from "@workspace/api-zod";
import { eq, and, isNull, inArray, sql } from "drizzle-orm";
import { getConnector } from "../../lore/serviceConnector.js";
import { h } from "../../middlewares/asyncHandler.js";
import { type AuthedRequest, getFreshToken } from "./auth.js";
import { enqueueRecordingEmbeds } from "../../lore/embed-resolution.js";
import { bandcampFridayInfo } from "../../lore/support-ladder.js";

const router: IRouter = Router();

/** Max MBIDs per batch keep-status check. */
const KEEP_BATCH_MAX = 50;

/** Client context that is useful for analytics but cannot attribute a play. */
const KEEP_CONTEXT_KEYS = ["surface", "entryPoint"] as const;

type KeepSpinProvenance = {
  stationSlug: string;
  stationName: string;
  showName: string | null;
  djName: string | null;
  pickerHandle: string | null;
  pickerName: string | null;
  playedAt: Date;
};

/**
 * POST /api/me/support-holds/:mbid — idempotently hold a recording for the
 * canonical next/current Bandcamp Friday. It deliberately does not touch
 * pending_keeps, which only represents unresolved spins.
 */
router.post(
  "/me/support-holds/:mbid",
  h(async (req, res) => {
    const user = (req as AuthedRequest).loreUser;
    const mbid = typeof req.params.mbid === "string" ? req.params.mbid.trim() : "";
    const parsed = HoldRecordingSupportParams.safeParse(req.params);
    if (!parsed.success || !mbid) return res.status(400).json({ error: "mbid is required" });
    const [recording] = await db
      .select({ mbid: recordingsTable.mbid })
      .from(recordingsTable)
      .where(eq(recordingsTable.mbid, mbid))
      .limit(1);
    if (!recording) return res.status(404).json({ error: "Recording not found" });

    const friday = bandcampFridayInfo();
    const [createdHold] = await db
      .insert(supportHoldsTable)
      .values({
        userId: user.id,
        recordingMbid: recording.mbid,
        bandcampFridayDate: friday.date,
      })
      .onConflictDoNothing()
      .returning();
    const [persistedHold] = createdHold
      ? [createdHold]
      : await db
          .select({ heldAt: supportHoldsTable.heldAt })
          .from(supportHoldsTable)
          .where(
            and(
              eq(supportHoldsTable.userId, user.id),
              eq(supportHoldsTable.recordingMbid, recording.mbid),
              eq(supportHoldsTable.bandcampFridayDate, friday.date),
            ),
          )
          .limit(1);

    return res.json(HoldRecordingSupportResponse.parse({
      mbid: recording.mbid,
      bandcampFridayDate: friday.date,
      held: true,
      heldAt: persistedHold?.heldAt.toISOString() ?? null,
    }));
  }),
);

/** DELETE is idempotent and only removes the canonical Friday hold. */
router.delete(
  "/me/support-holds/:mbid",
  h(async (req, res) => {
    const user = (req as AuthedRequest).loreUser;
    const mbid = typeof req.params.mbid === "string" ? req.params.mbid.trim() : "";
    const parsed = UnholdRecordingSupportParams.safeParse(req.params);
    if (!parsed.success || !mbid) return res.status(400).json({ error: "mbid is required" });
    const friday = bandcampFridayInfo();
    await db
      .delete(supportHoldsTable)
      .where(
        and(
          eq(supportHoldsTable.userId, user.id),
          eq(supportHoldsTable.recordingMbid, mbid),
          eq(supportHoldsTable.bandcampFridayDate, friday.date),
        ),
      );
    return res.json(UnholdRecordingSupportResponse.parse({
      mbid,
      bandcampFridayDate: friday.date,
      held: false,
      heldAt: null,
    }));
  }),
);

/**
 * Build keep provenance from the only trusted attribution source: the spin
 * row and its station/show/picker joins. Client input is limited to the two
 * non-attribution context fields used for analytics.
 *
 * Exported for the legacy provenance backfill so new writes and repairs
 * cannot drift into separate policies.
 */
export function buildKeepProvenance(
  override: unknown,
  spin: KeepSpinProvenance | null,
): LibraryItemProvenance {
  const provenance: LibraryItemProvenance = { kind: "keep" };

  if (override && typeof override === "object" && !Array.isArray(override)) {
    const input = override as Record<string, unknown>;
    for (const key of KEEP_CONTEXT_KEYS) {
      if (Object.prototype.hasOwnProperty.call(input, key)) {
        provenance[key] = input[key];
      }
    }
  }

  if (spin) {
    provenance.stationSlug = spin.stationSlug;
    provenance.stationName = spin.stationName;
    if (spin.showName != null) provenance.showName = spin.showName;
    if (spin.djName != null) provenance.djName = spin.djName;
    if (spin.pickerHandle != null) provenance.pickerHandle = spin.pickerHandle;
    if (spin.pickerName != null) provenance.pickerName = spin.pickerName;
    provenance.playedAt = spin.playedAt.toISOString();
  }

  return provenance;
}

// ---------------------------------------------------------------------------
// Keep endpoints
// ---------------------------------------------------------------------------

/**
 * POST /api/me/keep — upsert a recording into library_items and optionally
 * mirror to enabled streaming services.
 * Body: { mbid: string, provenance?: object }
 *    OR { spinId: number, provenance?: object }  ← unresolved-track path
 */
router.post(
  "/me/keep",
  h(async (req, res) => {
    const user = (req as AuthedRequest).loreUser;
    const {
      mbid,
      spinId,
      provenance: provenanceOverride,
    } = req.body as {
      mbid?: string;
      spinId?: number;
      provenance?: Partial<LibraryItemProvenance>;
    };

    // ── Spin-based save (unresolved or not-yet-resolved track) ────────────────
    if (!mbid && spinId != null) {
      const [spin] = await db
        .select({
          id: spinsTable.id,
          mbid: spinsTable.mbid,
          playedAt: spinsTable.playedAt,
          stationSlug: stationsTable.slug,
          stationName: stationsTable.name,
          showName: showsTable.name,
          djName: showsTable.djName,
          pickerHandle: pickersTable.handle,
          pickerName: pickersTable.name,
        })
        .from(spinsTable)
        .innerJoin(stationsTable, eq(spinsTable.stationId, stationsTable.id))
        .leftJoin(showsTable, eq(spinsTable.showId, showsTable.id))
        .leftJoin(pickersTable, eq(showsTable.pickerId, pickersTable.id))
        .where(eq(spinsTable.id, spinId))
        .limit(1);

      if (!spin) return res.status(404).json({ error: "Spin not found" });

      // If the spin already resolved, also write to library_items.
      let promotedAt: Date | null = null;
      if (spin.mbid) {
        const provenance = buildKeepProvenance(provenanceOverride, spin);
        await db
          .insert(libraryItemsTable)
          .values({
            userId: user.id,
            mbid: spin.mbid,
            provenance,
            spinId: spin.id,
            addedAt: new Date(),
          })
          .onConflictDoUpdate({
            target: [libraryItemsTable.userId, libraryItemsTable.mbid],
            // Explicit keep restores a previously removed row (removedAt: null).
            set: { provenance, spinId: spin.id, addedAt: new Date(), removedAt: null },
          });
        promotedAt = new Date();
      }

      await db
        .insert(pendingKeepsTable)
        .values({ userId: user.id, spinId: spin.id, promotedAt })
        .onConflictDoUpdate({
          target: [pendingKeepsTable.userId, pendingKeepsTable.spinId],
          set: { promotedAt },
        });

      // Show the recovery hint once — when this keep brings the total to exactly 3.
      const [spinLibCount] = await db
        .select({ n: sql<number>`count(*)::int` })
        .from(libraryItemsTable)
        .where(eq(libraryItemsTable.userId, user.id));
      const [spinPendingCount] = await db
        .select({ n: sql<number>`count(*)::int` })
        .from(pendingKeepsTable)
        .where(
          and(
            eq(pendingKeepsTable.userId, user.id),
            isNull(pendingKeepsTable.promotedAt),
          ),
        );
      const showRecoveryHint =
        (spinLibCount?.n ?? 0) + (spinPendingCount?.n ?? 0) === 3;

      return res.json({
        keptToLore: promotedAt != null,
        pendingKept: true,
        mirrors: [],
        showRecoveryHint,
      });
    }

    // ── MBID-based (resolved) keep ────────────────────────────────────────────
    if (!mbid || typeof mbid !== "string") {
      return res.status(400).json({ error: "mbid or spinId is required" });
    }

    // The recording must already be on the spine.
    const [recording] = await db
      .select()
      .from(recordingsTable)
      .where(eq(recordingsTable.mbid, mbid))
      .limit(1);

    if (!recording) {
      return res.status(404).json({ error: "Recording not on the spine" });
    }

    // When the client keeps a resolved track off a live play it can pass the
    // spin id alongside the mbid. Only store the link when the spin actually
    // resolved to this mbid — never persist mismatched provenance.
    let keepSpin: KeepSpinProvenance | null = null;
    if (spinId != null) {
      const [s] = await db
        .select({
          id: spinsTable.id,
          playedAt: spinsTable.playedAt,
          stationSlug: stationsTable.slug,
          stationName: stationsTable.name,
          showName: showsTable.name,
          djName: showsTable.djName,
          pickerHandle: pickersTable.handle,
          pickerName: pickersTable.name,
        })
        .from(spinsTable)
        .innerJoin(stationsTable, eq(spinsTable.stationId, stationsTable.id))
        .leftJoin(showsTable, eq(spinsTable.showId, showsTable.id))
        .leftJoin(pickersTable, eq(showsTable.pickerId, pickersTable.id))
        .where(and(eq(spinsTable.id, spinId), eq(spinsTable.mbid, mbid)))
        .limit(1);
      if (s) {
        keepSpin = {
          stationSlug: s.stationSlug,
          stationName: s.stationName,
          showName: s.showName,
          djName: s.djName,
          pickerHandle: s.pickerHandle,
          pickerName: s.pickerName,
          playedAt: s.playedAt,
        };
      }
    }
    const provenance = buildKeepProvenance(provenanceOverride, keepSpin);
    const keepSpinId = keepSpin ? spinId! : null;

    await db
      .insert(libraryItemsTable)
      .values({
        userId: user.id,
        mbid,
        provenance,
        spinId: keepSpinId,
        addedAt: new Date(),
      })
      .onConflictDoUpdate({
        target: [libraryItemsTable.userId, libraryItemsTable.mbid],
        set: {
          provenance,
          addedAt: new Date(),
          // Explicit keep restores a previously removed row.
          removedAt: null,
          ...(keepSpinId != null ? { spinId: keepSpinId } : {}),
        },
      });

    // A listener explicitly kept this recording: lift it ahead of cold-tail
    // work, but do not make the Keep request wait for any provider network IO.
    void enqueueRecordingEmbeds(mbid, { priority: 0 }).catch((err) =>
      console.warn("[lore] keep embed enqueue failed", err),
    );

    // Mirror to enabled service connectors.
    const enabledTargets = await db
      .select()
      .from(keepTargetsTable)
      .where(
        and(
          eq(keepTargetsTable.userId, user.id),
          eq(keepTargetsTable.enabled, true),
        ),
      );

    const mirrors: Array<{ service: string; ok: boolean; linkOut?: string }> =
      [];

    for (const target of enabledTargets) {
      const [conn] = await db
        .select()
        .from(serviceConnectionsTable)
        .where(
          and(
            eq(serviceConnectionsTable.userId, user.id),
            eq(serviceConnectionsTable.service, target.service),
          ),
        )
        .limit(1);

      if (!conn) {
        mirrors.push({ service: target.service, ok: false });
        continue;
      }

      if (!conn.canWrite) {
        const q = encodeURIComponent(`${recording.artist} ${recording.title}`);
        mirrors.push({
          service: target.service,
          ok: false,
          linkOut: `https://open.spotify.com/search/${q}`,
        });
        continue;
      }

      const accessToken = await getFreshToken(conn);
      if (!accessToken) {
        mirrors.push({ service: target.service, ok: false });
        continue;
      }

      const connector = getConnector(target.service);
      if (!connector) {
        mirrors.push({ service: target.service, ok: false });
        continue;
      }

      const result = await connector.addToLibrary(accessToken, recording);
      mirrors.push({ service: target.service, ...result });
    }

    // Show the recovery hint once — when this keep brings the total to exactly 3.
    const [libCount] = await db
      .select({ n: sql<number>`count(*)::int` })
      .from(libraryItemsTable)
      .where(eq(libraryItemsTable.userId, user.id));
    const [pendingCount] = await db
      .select({ n: sql<number>`count(*)::int` })
      .from(pendingKeepsTable)
      .where(
        and(
          eq(pendingKeepsTable.userId, user.id),
          isNull(pendingKeepsTable.promotedAt),
        ),
      );
    const showRecoveryHint = (libCount?.n ?? 0) + (pendingCount?.n ?? 0) === 3;

    return res.json({ keptToLore: true, mirrors, showRecoveryHint });
  }),
);

/**
 * DELETE /api/me/keep/spin/:spinId — remove a spin-based save.
 * Deletes from pending_keeps; if the spin resolved, also removes library_items.
 * Must be registered before DELETE /me/keep/:mbid or "spin" matches :mbid.
 */
router.delete(
  "/me/keep/spin/:spinId",
  h(async (req, res) => {
    const user = (req as AuthedRequest).loreUser;
    const spinId = parseInt(
      typeof req.params.spinId === "string" ? req.params.spinId : "",
      10,
    );
    if (isNaN(spinId)) return res.status(400).json({ error: "invalid spinId" });

    await db
      .delete(pendingKeepsTable)
      .where(
        and(
          eq(pendingKeepsTable.userId, user.id),
          eq(pendingKeepsTable.spinId, spinId),
        ),
      );

    // If the spin has an MBID, clean up library_items too.
    const [spin] = await db
      .select({ mbid: spinsTable.mbid })
      .from(spinsTable)
      .where(eq(spinsTable.id, spinId))
      .limit(1);

    if (spin?.mbid) {
      await db
        .delete(libraryItemsTable)
        .where(
          and(
            eq(libraryItemsTable.userId, user.id),
            eq(libraryItemsTable.mbid, spin.mbid),
          ),
        );
    }

    return res.status(204).end();
  }),
);

/**
 * GET /api/me/keep/pending-status?spinIds=1,2,3 — batch spin save-state check.
 * Returns two sets:
 *   savedSpinIds  — spin was saved AND promoted to library_items (resolved)
 *   pendingSpinIds — spin was saved but not yet resolved to an MBID
 */
router.get(
  "/me/keep/pending-status",
  h(async (req, res) => {
    const user = (req as AuthedRequest).loreUser;
    const rawIds =
      typeof req.query.spinIds === "string" ? req.query.spinIds : "";
    const spinIds = rawIds
      .split(",")
      .map((s) => parseInt(s.trim(), 10))
      .filter((n) => !isNaN(n) && n > 0)
      .slice(0, 50);

    if (spinIds.length === 0)
      return res.json({ savedSpinIds: [], pendingSpinIds: [] });

    const rows = await db
      .select({
        spinId: pendingKeepsTable.spinId,
        promotedAt: pendingKeepsTable.promotedAt,
      })
      .from(pendingKeepsTable)
      .where(
        and(
          eq(pendingKeepsTable.userId, user.id),
          inArray(pendingKeepsTable.spinId, spinIds),
        ),
      );

    const savedSpinIds = rows
      .filter((r) => r.promotedAt != null)
      .map((r) => r.spinId);
    const pendingSpinIds = rows
      .filter((r) => r.promotedAt == null)
      .map((r) => r.spinId);

    return res.json({ savedSpinIds, pendingSpinIds });
  }),
);

/**
 * DELETE /api/me/keep/:mbid — remove a recording from library_items only.
 * Never touches the streaming service library.
 */
router.delete(
  "/me/keep/:mbid",
  h(async (req, res) => {
    const user = (req as AuthedRequest).loreUser;
    const mbid = typeof req.params.mbid === "string" ? req.params.mbid : "";
    if (!mbid) return res.status(400).json({ error: "mbid is required" });

    await db
      .delete(libraryItemsTable)
      .where(
        and(
          eq(libraryItemsTable.userId, user.id),
          eq(libraryItemsTable.mbid, mbid),
        ),
      );

    return res.status(204).end();
  }),
);

/**
 * GET /api/me/keep/status?mbids=a,b,c — batch presence check.
 * Returns the subset of the given MBIDs that the user has kept.
 * Pattern mirrors GET /picks/contains.
 */
router.get(
  "/me/keep/status",
  h(async (req, res) => {
    const user = (req as AuthedRequest).loreUser;
    const raw =
      typeof req.query["mbids"] === "string" ? req.query["mbids"] : "";
    if (!raw) return res.status(400).json({ error: "mbids is required" });

    const mbids = [
      ...new Set(
        raw
          .split(",")
          .map((s) => s.trim())
          .filter(Boolean),
      ),
    ].slice(0, KEEP_BATCH_MAX);

    if (mbids.length === 0) return res.json({ kept: [] });

    const rows = await db
      .select({ mbid: libraryItemsTable.mbid })
      .from(libraryItemsTable)
      .where(
        and(
          eq(libraryItemsTable.userId, user.id),
          inArray(libraryItemsTable.mbid, mbids),
        ),
      );

    return res.json({ kept: rows.map((r) => r.mbid) });
  }),
);

export default router;