# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /me/support-holds/{mbid}:
    post:
      operationId: holdRecordingSupport
      tags: [me]
      summary: Hold a recording until the canonical Bandcamp Friday
      parameters:
        - name: mbid
          in: path
          required: true
          schema:
            type: string
            minLength: 1
      responses:
        "200":
          description: Hold state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupportHoldResponse"
        "404":
          description: Recording not on the spine
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      operationId: unholdRecordingSupport
      tags: [me]
      summary: Remove the current Bandcamp Friday hold
      parameters:
        - name: mbid
          in: path
          required: true
          schema:
            type: string
            minLength: 1
      responses:
        "200":
          description: Hold state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupportHoldResponse"
...
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    SupportHoldResponse:
      type: object
      properties:
        mbid:
          type: string
        bandcampFridayDate:
          type: string
        held:
          type: boolean
        heldAt:
          type: ["string", "null"]
      required: [mbid, bandcampFridayDate, held, heldAt]
...